refactor(api): put the services into seperate folder by convention
