+++ libs/api-implementation/orchestration/src/lib/orchestration.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { UtilsModule } from '../../../util/src/lib/util.module';
-import { VMOrchestrationService } from './orchestration.service';
+import { VMOrchestrationService } from './services/orchestration.service';
 import { VMOrchestrationController } from './orchestration.controller';
-import { Compute } from './compute.service';
+import { Compute } from './services/compute.service';
 
 @Module({
   imports: [UtilsModule],
